requirements.txt
streamlit==1.32.0
pandas==2.2.0
mysql-connector-python==8.3.0
plotly==5.18.0
textblob==0.17.1
wordcloud==1.9.3
matplotlib==3.8.2
python-dotenv==1.0.0 